Keep refresh appearance tests in existing suite
